requirements.txt
requests
sqlalchemy
straight.plugin
jinja2
python-geoip-geolite2
lxml
iptools
